# golang/internal/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: wow-converter
  version: "1.1.0"
  description: |
    This server reads a World of Warcraft installation and hands you the files.
    Model, icon, and map calls convert those files into the formats Warcraft III uses.
    Sound calls do not convert anything. They return the original audio.
    GET /api/sound/transcripts is a separate list of spoken lines for those files.

    There is no login. Anyone who can open the port can call every endpoint.
    A server you run yourself listens on 127.0.0.1 port 3001 unless HOST or PORT was changed.
    A public shared server listens on every interface and refuses calls that would
    change the game install, delete files, or build a map.

    ## How a call looks

    Send JSON as application/json. File paths inside the game use forward slashes.
    When a path is part of the URL, it is the remainder of that URL and it contains slashes.
    Percent-encoding those slashes is accepted.

    Lists are not paginated. Pass search to keep the body small. An empty search is the whole catalog.

    A 200 that is audio, PNG, or a zip is the file itself, not JSON. A failure on that same URL is JSON.
    Read the status code before deciding how to read the body. 304 means the body you already have is still current, and this response has no body.

    Opening an edition, and the first list after that, can take minutes. A client that gives up after a few seconds will fail a call that is still working.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/export/character:
    post:
      tags: [Character export]
      summary: Convert one model to Warcraft III
      description: |
        Use this to turn one creature, player, or item into a Warcraft III model plus the images it needs.
        The 200 body is a job object, not the model. Read id from it and poll GET /api/export/character/status/{jobId}.
        If cascLoaded is false, this returns 500. Check status before queueing.

        base says which model. type wowhead takes a public Wowhead page URL, such as a creature or item page.
        type displayID takes that appearance's numeric id as a string.
        type local takes a path from the browse list, such as a skin localRef. It must be relative and must not contain .. .

        outputFileName is the file name to write, without an extension and without .. . The server adds .mdx or .mdl.
        format mdx is the binary model. format mdl is the same model as text, which is easier to diff.
        formatVersion 1000 is the current Warcraft III model version. 800 is the older one. Omit it for 1000.

        isBrowse writes a preview into a separate folder served at /api/browse-assets/.
        A desktop server deletes that folder on startup. Leave isBrowse false to keep the files under /api/assets/.
        On a shared server, and for every preview, a repeat of the same request returns the existing job instead of starting another,
        and the written name gets a stable suffix so two different requests do not overwrite each other.

        optimization is required, but an empty object is valid. Omitted flags then use the defaults below.
        The optimization object shrinks the model. By default the server sorts animations and drops unused vertices, nodes, materials, and images.
        allMaterialsUnshaded defaults to false. Set it true to strip lighting from every surface.
        maxTextureSize, when set, shrinks images so neither side is larger than 256, 512, or 1024.
      operationId: exportCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterExportRequest'
            example:
              character:
                base:
                  type: wowhead
                  value: "https://www.wowhead.com/npc=36597/the-lich-king"
                inGameMovespeed: 270
              outputFileName: lich-king
              optimization: {}
              format: mdx
      responses:
        '200':
          description: The queued job, or the existing job when the request was deduplicated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterJob'
        '400':
          description: Field errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '413':
          $ref: '#/components/responses/Error'
        '500':
          description: Game data is not ready, or the export failed before a job was queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/export/character/status/{jobId}:
    get:
      tags: [Character export]
      summary: Read a model-export job
      description: |
        Poll this until status is done, failed, or cancelled. pending means it is waiting. processing means it is running.
        While it is running, logs is the text written so far.
        When it is done, result lists the files. Download them with POST /api/download or GET /api/assets/{path}.
        The id stops working 5 minutes after the job finishes.
      operationId: getCharacterExportStatus
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterJob'
        '404':
          description: Unknown id, or the job has already expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/export/character/check-local-file:
    get:
      tags: [Character export]
      summary: Check a model path before you export it
      description: |
        Call this before a local export so a typo fails here instead of inside a job.
        ok is true when that path is a model in the open edition.
        When it is not, similarFiles lists close paths. A missing or unsafe path is ok false and an empty list, not an error.
      operationId: checkLocalFile
      parameters:
        - name: localPath
          in: query
          schema: { type: string }
          description: Path from the browse list, the same string you would send as a local model source.
      responses:
        '200':
          description: Whether the path resolves.
          content:
            application/json:
              schema:
                type: object
                required: [ok, similarFiles]
                properties:
                  ok: { type: boolean }
                  similarFiles:
                    type: array
                    items: { type: string }
  /api/export/character/clean:
    post:
      tags: [Character export]
      summary: Delete every exported file
      description: |
        Wipes both the kept exports and the preview folder, and clears the recent list.
        A shared server returns 400. A desktop server also returns an error when exports were configured to live somewhere other than the default folder.
        No body.
      operationId: cleanExports
      responses:
        '200':
          description: Exports were deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    ValidationError:
      type: object
      required: [error, issues]
      properties:
        error: { type: string, example: Invalid request body }
        issues:
          type: array
          items: { type: string }
...
    ModelRef:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [local, wowhead, displayID]
          description: |
            local is a path from the browse list. wowhead is a Wowhead page URL for a creature or item.
            displayID is that appearance's numeric id, as text.
        value: { type: string, description: The path, URL, or id, depending on type. }
    AttachItem:
      type: object
      required: [path]
      properties:
        path:
          $ref: '#/components/schemas/ModelRef'
        scale: { type: number }
    Mount:
      type: object
      required: [path]
      properties:
        path:
          $ref: '#/components/schemas/ModelRef'
        scale: { type: number }
        seatOffset:
          type: array
          minItems: 3
          maxItems: 3
          items: { type: number }
          description: X, Y, Z offset of the rider.
        animation: { type: string }
    Character:
      type: object
      required: [base, inGameMovespeed]
      properties:
        base:
          $ref: '#/components/schemas/ModelRef'
        inGameMovespeed:
          type: number
          description: |
            How fast this creature moves in the source game, in that game's speed units.
            The walk animation is timed to match it.
        attackTag:
          type: string
          enum: ['', Auto, 1H, 2H, 2HL, Unarmed, Bow, Rifle, Thrown]
          description: Which weapon animation set to keep. Empty or Auto lets the server choose. 1H is one-handed, 2H two-handed, 2HL a long two-hander.
        size:
          type: string
          enum: [small, medium, large, hero, semi-giant, giant]
          description: A named size preset for how big the model is in Warcraft III.
        scale: { type: number, description: Extra multiplier on top of size. 1 leaves the preset unchanged. }
        keepCinematic: { type: boolean }
        noDecay: { type: boolean }
        forceSheathed: { type: boolean }
        withCollision: { type: boolean }
        particlesDensity: { type: number }
        portraitCameraSequenceName: { type: string }
        attachItems:
          type: object
          description: |
            Extra models stuck to sockets on the body, such as a weapon or shield.
            Each key is a socket number as text. Common ones are 0 shield, 1 right hand, 2 left hand.
          additionalProperties:
            $ref: '#/components/schemas/AttachItem'
        mount:
          $ref: '#/components/schemas/Mount'
    CharacterOptimization:
      type: object
      properties:
        sortSequences: { type: boolean, default: true }
        allMaterialsUnshaded: { type: boolean, default: false }
        removeUnusedVertices: { type: boolean, default: true }
        removeUnusedNodes: { type: boolean, default: true }
        removeUnusedMaterialsTextures: { type: boolean, default: true }
        maxTextureSize: { type: string, enum: ['256', '512', '1024'] }
    CharacterExportRequest:
      type: object
      required: [character, outputFileName, optimization, format]
      properties:
        character:
          $ref: '#/components/schemas/Character'
        outputFileName:
          type: string
          description: File name to write, without an extension. Relative, and it must not contain .. .
        optimization:
          $ref: '#/components/schemas/CharacterOptimization'
        format:
          type: string
          enum: [mdx, mdl]
          description: mdx is the binary Warcraft III model. mdl is the same model as text.
        formatVersion:
          type: string
          enum: ['800', '1000']
          default: '1000'
          description: 1000 is the current model version. 800 is the older one. Omit this field for 1000.
        isBrowse:
          type: boolean
          description: |
            True writes a preview into a separate folder that a desktop server deletes on startup, and skips the recent list.
            False keeps the files.
        skinId:
          type: string
          description: Which texture variant to export. The id from GET /api/browse/model-skins. Omit it for the default look.
    ExportedFile:
      type: object
      properties:
        path: { type: string, description: Relative path to pass to /api/download or /api/assets. }
        size: { type: integer, format: int64, description: Size in bytes. }
    ModelStats:
      type: object
      description: |
        How big the written model is. Keys are counts such as vertices, faces, bones, and sequences.
        formatVersion is 800 or 1000, the Warcraft III model version that was written.
      additionalProperties: { type: integer }
    CharacterExportResult:
      type: object
      properties:
        exportedModels:
          type: array
          items:
            $ref: '#/components/schemas/ExportedFile'
        exportedTextures:
          type: array
          items:
            $ref: '#/components/schemas/ExportedFile'
        modelStats:
          $ref: '#/components/schemas/ModelStats'
        outputDirectory:
          type: string
          description: Folder on this machine. Left out on a shared server, which does not reveal local paths.
        versionId: { type: string, description: Same as the job id. }
    CharacterJob:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [pending, processing, done, failed, cancelled] }
        position: { type: integer, description: Place in the queue while pending. }
        result:
          $ref: '#/components/schemas/CharacterExportResult'
        error: { type: string }
        submittedAt: { type: integer, format: int64 }
        startedAt: { type: integer, format: int64 }
        finishedAt: { type: integer, format: int64 }
        logs:
          type: array
          description: Present only while status is processing.
          items: { type: string }
...